+++ test/samples/deps-project/src/clj/hello-clj/world.clj
@@ -1,0 +1,8 @@
+(ns hello-clj.world
+  (:require
+   ;; This is a comment
+   [clojure.string :as str]
+   ;; Hello world
+   [clojure.math :as math])
+  (:import
+   (java.util.time Instant ZonedDateTime)))