+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## main (unreleased)
 
-- Add a dedicated mode for editing Joker code. (`clojure-ts-joker-mode`)
+- Add a dedicated mode for editing Joker code. (`clojure-ts-joker-mode`).
+- [#113](https://github.com/clojure-emacs/clojure-ts-mode/pull/113): Fix non-working refactoring commands for Emacs-30.
 
 ## 0.5.1 (2025-06-17)