+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [#116](https://github.com/clojure-emacs/clojure-ts-mode/pull/116): Extend built-in completion to complete all imported symbols from an `ns`
   form.
 - Add documentation and bug reporting commands from `clojure-mode`.
+- Add some ns manipulation functions from `clojure-mode`.
 
 ## 0.5.1 (2025-06-17)